mvp progress update
* refactor pt. 1

* initial test commit new laptop

* refactor repository layer

* refactor WiP

* create database tables and data

* refactor 60% done

* add buying and selling logic

* Next day logic

* start nextDayLogic

* nextDayLogic backend logic done

* debugging session

* bugfix

* final bugfix - no more errors

* implement buying and selling items

* actions fix

* update actions workflow and dependencies

* tables in actions workflow updated

* vite update

* Inventory Refactor

* inventoryRepo WiP

* start inventoryspace movement to player

* inventorySpace refactor WiP

* implement saveInventory, finish inventorySpace refactore

* rollup dependency update

* create and refactor item images and databases

* Create Inventory Entity and implement HighScore Logic

* implement inventory converter logic and refactor

* implement highscore backend logic

* fix package.json

* Implement Merchant Selection and nextDay button

* backend Done

* refactor from oop to functional for react to work

* merchant selection working

* create nextDay button

* get latest updates into dev

* backend Done

* refactor from oop to functional for react to work

* merchant selection working

* create nextDay button

* item shows in focusWindow when clicked

* correctly show information of focusItem

* convert backend currency from double to int

* Get Newest data to develop

* backend Done

* refactor from oop to functional for react to work

* merchant selection working

* create nextDay button

* item shows in focusWindow when clicked

* correctly show information of focusItem

* convert backend currency from double to int

* test commit

* revert test commit

* frontend currency showing correct

* delete interface

* implement buy button

* write unit tests for playerService

* implement sell button and inventory management logic

* change menu appearance

* move nextDayButton to Menu

* show correnct item prices in focusItemMenu

* merchantInventory is updated when buying item

* fix bug when merchant doesn't have item

* disable button when condition is met

---------

+++ src/main/kotlin/com/projects/inGameMarketplace/playerService/PlayerEntity.kt
@@ -7,5 +7,5 @@ data class PlayerEntity(
     val playerName: String,
     val inventorySpace: Int,
     val day: Int,
-    val money: Double,
+    val money: Int,
 )
